adding initial ROI

+++ modules/unsupported/process-raster/src/main/java/org/geotools/process/raster/changematrix/ChangeMatrixElement.java
@@ -35,14 +35,14 @@ public class ChangeMatrixElement implements Comparable<ChangeMatrixElement>{
     /**
      * The number of pixel that has been changed from class ref to class now
      */
-    private int pixels;
+    private long pixels;
     
     /**
      * @param ref
      * @param now
      * @param pixels
      */
-    public ChangeMatrixElement(int ref, int now, int pixels) {
+    public ChangeMatrixElement(int ref, int now, long pixels) {
         this.ref = ref;
         this.now = now;
         this.pixels = pixels;
@@ -65,7 +65,7 @@ public class ChangeMatrixElement implements Comparable<ChangeMatrixElement>{
     /**
      * @return the pixels
      */
-    public int getPixels() {
+    public long getPixels() {
         return pixels;
     }
